remove more cobweb traits
